+++ docs/CHANGE_LOG.md
@@ -13,6 +13,9 @@
 #### **Documentation:**
 - Updated the documentation to show the right properties for the objects, and add missing documentation for the binding attribute.
 
+#### **Fixes:**
+- Fixed a huge fault that prevented all other binding properties (other than the "json") to be ignored when set in DOM attribute "data-jhson-js".
+
 <br>